Refactor toolbar menu code
At the moment all code related to the menu resides in menu.js with only a single entry point function exported. This makes it hard to test as well as could potentially grow into a large class.
- moved everything under a menu directory for ease of navigation
- split buttons and dropdown code since they have pretty distinct patterns and don't share common code
