+++ lib/plugins.js
@@ -3,7 +3,7 @@ import customInputRules from "./plugins/custom-input-rules";
 import customKeymap from "./plugins/custom-keymap";
 import editorGovspeakClass from "./plugins/editor-govspeak-class";
 import linkTooltip from "./plugins/link-tooltip";
-import menu from "./plugins/menu";
+import menu from "./plugins/menu/menu";
 
 export default (options) => {
   const plugins = [
