+++ lib/plugins/menu/menu-plugin-view.js
